# swagger.yaml
swagger: "2.0"
info:
  title: GameHub API
  version: "1.0"
  description: API REST para gestionar productos y carrito de compras de videojuegos.
host: 127.0.0.1:5000
basePath: /
schemes:
  - http
...
paths:
...
  /carrito/agregar:
    post:
      tags:
        - Carrito

      summary: Agregar producto al carrito

      parameters:
        - in: body
          name: body

          schema:
            type: object

            properties:
              id:
                type: integer
                example: 1

      responses:
        200:
          description: Producto agregado correctamente

        404:
          description: Producto no encontrado
...